Created the AddCardDialog interface for adding a card within a book. Added Edit Book Name functionality selection to Book's Menu.

+++ app/src/main/java/edu/ucdenver/sweanyn/flashcardfinal/BookDao.java
@@ -16,7 +16,7 @@ public interface BookDao {
     List<Book> getAll();
 
     @Query("SELECT * FROM books WHERE id = :bookId")
-    Book getById(int bookId);
+    Book getById(long bookId);
 
     @Query("SELECT * FROM books WHERE book_name = :bookName LIMIT 1")
     Book findByName(String bookName);
